+++ include/boost/astronomy/coordinate/coord_sys/supergalactic.hpp
@@ -9,9 +9,9 @@
 #ifndef BOOST_ASTRONOMY_COORDINATE_SUPERGALACTIC_HPP
 #define BOOST_ASTRONOMY_COORDINATE_SUPERGALACTIC_HPP
 
-#include <boost/astronomy/coordinate/base_frame.hpp>
-#include <boost/astronomy/coordinate/spherical_representation.hpp>
-#include <boost/astronomy/coordinate/spherical_coslat_differential.hpp>
+#include <boost/astronomy/coordinate/reference_frames/base_frame.hpp>
+#include <boost/astronomy/coordinate/rep/spherical_representation.hpp>
+#include <boost/astronomy/coordinate/differentials/spherical_coslat_differential.hpp>
 
 namespace boost { namespace astronomy { namespace coordinate {
