+++ include/boost/astronomy/coordinate/diff/spherical_coslat_differential.hpp
@@ -7,12 +7,13 @@
   file LICENSE_1_0.txt or copy at https://www.boost.org/LICENSE_1_0.txt)
 =============================================================================*/
 
-#ifndef BOOST_ASTRONOMY_COORDINATE_SPHERICAL_EQUATORIAL_DIFFERENTIAL_HPP
-#define BOOST_ASTRONOMY_COORDINATE_SPHERICAL_EQUATORIAL_DIFFERENTIAL_HPP
+#ifndef BOOST_ASTRONOMY_COORDINATE_SPHERICAL_COSLAT_DIFFERENTIAL_HPP
+#define BOOST_ASTRONOMY_COORDINATE_SPHERICAL_COSLAT_DIFFERENTIAL_HPP
 
 #include <tuple>
 #include <type_traits>
 
+#include <math.h>
 #include <boost/geometry/strategies/strategy_transform.hpp>
 #include <boost/geometry/core/cs.hpp>
 #include <boost/geometry/geometries/point.hpp>
@@ -25,16 +26,18 @@
 #include <boost/units/systems/si/plane_angle.hpp>
 
 #include <boost/astronomy/detail/is_base_template_of.hpp>
-#include <boost/astronomy/coordinate/base_differential.hpp>
-#include <boost/astronomy/coordinate/cartesian_differential.hpp>
+#include <boost/astronomy/coordinate/diff/base_differential.hpp>
+#include <boost/astronomy/coordinate/diff/cartesian_differential.hpp>
+#include <boost/astronomy/coordinate/diff/spherical_differential.hpp>
+
 
 namespace boost { namespace astronomy { namespace coordinate {
 
 namespace bu = boost::units;
 namespace bg = boost::geometry;
 
-//!Represents the differential in spherical equatorial representation
-//!Uses three components to represent a differential (dlatitude, dlongitude, ddistance)
+//!Represents the differential in spherical representation including cos(latitude) term
+//!Uses three components to represent a differential (dlatitude, dlongitude_coslat, ddistance)
 
 template
 <
@@ -43,8 +46,8 @@ template
     typename LonQuantity = bu::quantity<bu::si::plane_angle, CoordinateType>,
     typename DistQuantity = bu::quantity<bu::si::dimensionless, CoordinateType>
 >
-struct spherical_equatorial_differential : public base_differential
-    <3, bg::cs::spherical_equatorial<radian>, CoordinateType>
+struct spherical_coslat_differential : public base_differential
+    <3, geometry::cs::spherical<radian>, CoordinateType>
 {
     ///@cond INTERNAL
     BOOST_STATIC_ASSERT_MSG(
@@ -63,17 +66,17 @@ public:
     typedef DistQuantity quantity3;
 
     //default constructor no initialization
-    spherical_equatorial_differential() {}
+    spherical_coslat_differential() {}
 
     //!constructs object from provided components of differential
-    spherical_equatorial_differential
+    spherical_coslat_differential
     (
         LatQuantity const& dlat,
-        LonQuantity const& dlon,
+        LonQuantity const& dlon_coslat,
         DistQuantity const& ddistance
     )
     {
-        this->set_dlat_dlon_ddist(dlat, dlon, ddistance);
+        this->set_dlat_dlon_coslat_ddist(dlat, dlon_coslat, ddistance);
     }
 
     //!constructs object from boost::geometry::model::point object
@@ -83,7 +86,7 @@ public:
         typename OtherCoordinateSystem,
         typename OtherCoordinateType
     >
-    spherical_equatorial_differential
+    spherical_coslat_differential
     (
         bg::model::point
         <
@@ -99,36 +102,39 @@ public:
     }
 
     //copy constructor
-    spherical_equatorial_differential
+    spherical_coslat_differential
     (
-        spherical_equatorial_differential
+        spherical_coslat_differential
         <
             CoordinateType,
             LatQuantity,
             LonQuantity,
             DistQuantity
-        > const& object
+        > const& other
     )
     {
-        this->diff = object.get_differential();
+        this->diff = other.get_differential();
     }
 
-    //!constructs object from any type of differential
+    // !constructs object from any type of differential
     template <typename Differential>
-    spherical_equatorial_differential(Differential const& other)
+    spherical_coslat_differential(Differential const& other)
     {
         BOOST_STATIC_ASSERT_MSG((boost::astronomy::detail::is_base_template_of
             <boost::astronomy::coordinate::base_differential, Differential>::value),
             "No constructor found with given argument type");
 
-        auto temp = make_spherical_equatorial_differential(other);
-        bg::transform(temp.get_differential(), this->diff);
-
+        auto temp = make_spherical_differential(other);
+        temp.set_dlon(temp.get_dlon() * cos(static_cast<bu::quantity
+            <bu::si::plane_angle, CoordinateType>>(temp.get_dlat()).value()));
+        this->diff = temp.get_differential();
     }
-    //! returns the (dlat, dlon, ddistance) in the form of tuple
-    std::tuple<LatQuantity, LonQuantity, DistQuantity> get_dlat_dlon_ddist() const
+
+    //! returns the (dlat, dlon_coslat, ddistance) in the form of tuple
+    std::tuple<LatQuantity, LonQuantity, DistQuantity> get_dlat_dlon_coslat_ddist() const
     {
-        return std::make_tuple(this->get_dlat(), this->get_dlon(), this->get_ddist());
+        return std::make_tuple(this->get_dlat(), this->get_dlon_coslat(),
+            this->get_ddist());
     }
 
     //!returns the dlat component of differential
@@ -141,8 +147,8 @@ public:
             );
     }
 
-    //!returns the dlon component of differential
-    LonQuantity get_dlon() const
+    //!returns the dlon_coslat component of differential
+    LonQuantity get_dlon_coslat() const
     {
         return static_cast<LonQuantity>
             (
@@ -157,16 +163,16 @@ public:
         return DistQuantity::from_value(bg::get<2>(this->diff));
     }
 
-    //!set value of (dlat, dlon, ddistance) in current object
-    void set_dlat_dlon_ddist
+    //!set value of (dlat, dlon_coslat, ddistance) in current object
+    void set_dlat_dlon_coslat_ddist
     (
         LatQuantity const& dlat,
-        LonQuantity const& dlon,
+        LonQuantity const& dlon_coslat,
         DistQuantity const& ddistance
     )
     {
         this->set_dlat(dlat);
-        this->set_dlon(dlon);
+        this->set_dlon_coslat(dlon_coslat);
         this->set_ddist(ddistance);
     }
 
@@ -180,13 +186,14 @@ public:
             );
     }
 
-    //!set value of dlon component of differential
-    void set_dlon(LonQuantity const& dlon)
+    //!set value of dlon_coslat component of differential
+    void set_dlon_coslat(LonQuantity const& dlon_coslat)
     {
         bg::set<1>
             (
             this->diff,
-            static_cast<bu::quantity<bu::si::plane_angle, CoordinateType>>(dlon).value()
+            static_cast<bu::quantity<bu::si::plane_angle, CoordinateType>>
+                (dlon_coslat).value()
             );
     }
 
@@ -201,7 +208,7 @@ public:
     <
         typename Addend
     >
-    spherical_equatorial_differential
+    spherical_coslat_differential
     <
         CoordinateType,
         LatQuantity,
@@ -216,13 +223,13 @@ public:
 
         auto temp = cartesian1 + cartesian2;
 
-        spherical_equatorial_differential
+        spherical_coslat_differential
         <
             CoordinateType,
             LatQuantity,
             LonQuantity,
             DistQuantity
-        > result = make_spherical_equatorial_differential(temp);
+        > result = make_spherical_coslat_differential(temp);
 
         return result;
     }
@@ -235,10 +242,10 @@ public:
     auto operator *(OtherQuantity const& dt) const
     {
         
-        spherical_equatorial_differential
+        spherical_coslat_differential
             <CoordinateType, LatQuantity, LonQuantity, DistQuantity> temp(this->diff);
 
-        spherical_equatorial_differential
+        spherical_coslat_differential
         <
             CoordinateType,
             LatQuantity,
@@ -252,13 +259,13 @@ public:
         > product
         (
             temp.get_dlat(),
-            temp.get_dlon(),
+            temp.get_dlon_coslat(),
             temp.get_ddist() * dt
         );
 
         return product;
     }
-}; //spherical_equatorial_differential
+}; //spherical_coslat_differential
 
 //!constructs object from provided components of differential
 template
@@ -271,27 +278,27 @@ template
     typename Unit2,
     typename Unit3
 >
-spherical_equatorial_differential
+spherical_coslat_differential
 <
     CoordinateType,
     LatQuantity<Unit1, CoordinateType>,
     LonQuantity<Unit2, CoordinateType>,
     DistQuantity<Unit3, CoordinateType>
 >
-make_spherical_equatorial_differential
+make_spherical_coslat_differential
 (
     LatQuantity<Unit1, CoordinateType> const& dlat,
-    LonQuantity<Unit2, CoordinateType> const& dlon,
+    LonQuantity<Unit2, CoordinateType> const& dlon_coslat,
     DistQuantity<Unit3, CoordinateType> const& ddist
 )
 {
-    return spherical_equatorial_differential
+    return spherical_coslat_differential
         <
             CoordinateType,
             LatQuantity<Unit1, CoordinateType>,
             LonQuantity<Unit2, CoordinateType>,
             DistQuantity<Unit3, CoordinateType>
-        >(dlat, dlon, ddist);
+        >(dlat, dlon_coslat, ddist);
 }
 
 //!constructs object from provided components of differential with different units
@@ -306,16 +313,16 @@ template
     typename LonQuantity,
     typename DistQuantity
 >
-spherical_equatorial_differential
+spherical_coslat_differential
 <
     ReturnCoordinateType,
     ReturnLatQuantity,
     ReturnLonQuantity,
     ReturnDistQuantity
 >
-make_spherical_equatorial_differential
+make_spherical_coslat_differential
 (
-    spherical_equatorial_differential
+    spherical_coslat_differential
     <
         CoordinateType,
         LatQuantity,
@@ -324,9 +331,9 @@ make_spherical_equatorial_differential
     > const& other
 )
 {
-    return make_spherical_equatorial_differential(
+    return make_spherical_coslat_differential(
         static_cast<ReturnLatQuantity>(other.get_dlat()),
-        static_cast<ReturnLonQuantity>(other.get_dlon()),
+        static_cast<ReturnLonQuantity>(other.get_dlon_coslat()),
         static_cast<ReturnDistQuantity>(other.get_ddist())
     );
 }
@@ -339,16 +346,16 @@ template
     typename LonQuantity,
     typename DistQuantity
 >
-spherical_equatorial_differential
+spherical_coslat_differential
 <
     CoordinateType,
     LatQuantity,
     LonQuantity,
     DistQuantity
 >
-make_spherical_equatorial_differential
+make_spherical_coslat_differential
 (
-    spherical_equatorial_differential
+    spherical_coslat_differential
     <
         CoordinateType,
         LatQuantity,
@@ -357,7 +364,7 @@ make_spherical_equatorial_differential
     > const& other
 )
 {
-    return spherical_equatorial_differential
+    return spherical_coslat_differential
         <
             CoordinateType,
             LatQuantity,
@@ -377,14 +384,14 @@ template
     typename OtherCoordinateSystem,
     typename OtherCoordinateType
 >
-spherical_equatorial_differential
+spherical_coslat_differential
 <
     CoordinateType,
     LatQuantity,
     LonQuantity,
     DistQuantity
 >
-make_spherical_equatorial_differential
+make_spherical_coslat_differential
 (
     bg::model::point
     <
@@ -394,7 +401,7 @@ make_spherical_equatorial_differential
     > const& pointObject
 )
 {
-    return spherical_equatorial_differential
+    return spherical_coslat_differential
         <
             CoordinateType,
             LatQuantity,
@@ -408,42 +415,73 @@ template
 <
     typename OtherDifferential
 >
-auto make_spherical_equatorial_differential
+auto make_spherical_coslat_differential
 (
     OtherDifferential const& other
 )
-{
-    auto temp = make_cartesian_differential(other);
-    typedef decltype(temp) cartesian_type;
+{   
+    auto temp = make_spherical_differential(other);
+    typedef decltype(temp) spherical_type;
 
-    bg::model::point
+    temp.set_dlon(temp.get_dlon() * cos(static_cast<bu::quantity
+        <bu::si::plane_angle, typename spherical_type::type>>(temp.get_dlat()).value()));
+    
+    return spherical_coslat_differential
         <
-            typename cartesian_type::type,
-            3,
-            bg::cs::cartesian
-        > tempPoint;
-
-    bg::set<0>(tempPoint, temp.get_dx().value());
-    bg::set<1>(tempPoint, static_cast<typename cartesian_type::quantity1>
-        (temp.get_dy()).value());
-    bg::set<2>(tempPoint, static_cast<typename cartesian_type::quantity1>
-        (temp.get_dz()).value());
-
-    bg::model::point<typename cartesian_type::type, 3, bg::cs::spherical_equatorial
-        <radian>> result;
-    bg::transform(tempPoint, result);
-
-    return spherical_equatorial_differential
-        <
-            typename cartesian_type::type,
-            bu::quantity<bu::si::plane_angle, typename cartesian_type::type>,
-            bu::quantity<bu::si::plane_angle, typename cartesian_type::type>,
-            typename cartesian_type::quantity1
-        >(result);
+            typename spherical_type::type,
+            bu::quantity<bu::si::plane_angle, typename spherical_type::type>,
+            bu::quantity<bu::si::plane_angle, typename spherical_type::type>,
+            typename spherical_type::quantity3
+        >(temp.get_differential());
 }
 
-}}}//namespace boost::astronomy::coordinate
+//!constructs cartesian object from spherical_coslat_differential
+//function placed in this file to avoid circular dependency of header files
+template
+<
+    typename CoordinateType,
+    template <typename Unit1, typename CoordinateType_> class LatQuantity,
+    template <typename Unit2, typename CoordinateType_> class LonQuantity,
+    template <typename Unit3, typename CoordinateType_> class DistQuantity,
+    typename Unit1,
+    typename Unit2,
+    typename Unit3
+>
+auto
+make_cartesian_differential
+(
+    spherical_coslat_differential
+    <
+        CoordinateType,
+        LatQuantity<Unit1, CoordinateType>,
+        LonQuantity<Unit2, CoordinateType>,
+        DistQuantity<Unit3, CoordinateType>
+    > const& other
+)
+{
+    typedef spherical_coslat_differential
+    <
+        CoordinateType,
+        LatQuantity<Unit1, CoordinateType>,
+        LonQuantity<Unit2, CoordinateType>,
+        DistQuantity<Unit3, CoordinateType>
+    > spherical_type;
+    bg::model::point<typename spherical_type::type, 3, bg::cs::spherical<radian>> temp;
 
-#endif // !BOOST_ASTRONOMY_COORDINATE_SPHERICAL_EQUATORIAL_DIFFERENTIAL_HPP
+    temp = other.get_differential();
 
+    bg::set<1>(temp, bg::get<1>(temp) / cos(bg::get<0>(temp)));
+
+    return cartesian_differential
+    <
+        typename spherical_type::type,
+        typename spherical_type::quantity3,
+        typename spherical_type::quantity3,
+        typename spherical_type::quantity3
+    >(temp);
+}
+
+}}} //namespace boost::astronomy::coordinate
+
+#endif // !BOOST_ASTRONOMY_COORDINATE_SPHERICAL_COSLAT_DIFFERENTIAL_HPP
